added static casting

+++ approxPi.cpp
@@ -11,9 +11,9 @@
 #include <iostream>
 using namespace std;
 
-double approximate(int loopNum){
+double approximate(double loopNum){
   double out = 0;
-  int sign = 1;
+  double sign = 1;
 
   for(double i = 1.0; i < loopNum * 2; i += 2){
     out += (1/i) * sign;
@@ -29,7 +29,7 @@ int main(){
 
   cout << "\n*** Start Extra Credit ***\n" << endl;
 
-  cout << "PI = " << ((int)(approximate(152) * 100 + 0.5)) / 100.0 << endl;
+  cout << "PI = " << static_cast<double>(static_cast<int>(approximate(152) * 100 + 0.5)) / 100.0 << endl;
 
   cout << "\n*** End Extra Credit ***" << endl;
   return 0;
